requirements.txt
Flask
uwsgi
Pillow

# gunicorn if needed